refactor(launcher): decompose UI setup responsibilities
