+++ scripts/config/module_size_budget_baseline.json
@@ -1,11 +1,4 @@
 {
   "max_lines": 1200,
-  "exceptions": [
-    {
-      "path": "src/launchers/launcher_ui_setup.py",
-      "owner": "@dieterolson @D-sorganization/core-maintainers",
-      "reason": "Legacy launcher UI setup module (2105 lines, pending decomposition under #5922)",
-      "expires_on": "2026-07-31"
-    }
-  ]
+  "exceptions": []
 }
