+++ src/launchers/launcher_ui_setup.py
@@ -11,25 +11,19 @@ from __future__ import annotations
 import datetime
 from typing import TYPE_CHECKING, Any
 
-from PyQt6.QtCore import QSize, Qt, QTimer, pyqtSignal, QEvent, QObject
+from PyQt6.QtCore import Qt, pyqtSignal, QObject
 from PyQt6.QtGui import (
     QAction,
     QKeySequence,
     QShortcut,
-    QColor,
-    QMouseEvent,
-    QEnterEvent,
 )
 from PyQt6.QtWidgets import (
-    QButtonGroup,
     QCheckBox,
     QComboBox,
     QFrame,
-    QGraphicsDropShadowEffect,
     QGridLayout,
     QHBoxLayout,
     QLabel,
-    QMenuBar,
     QPlainTextEdit,
     QPushButton,
     QScrollArea,
@@ -44,16 +38,23 @@ from PyQt6.QtWidgets import (
 from src.launchers.launcher_constants import (
     TILE_SCALE_MAX,
     TILE_SCALE_MIN,
-    ViewMode,
 )
 from src.launchers.custom_title_bar import create_window_control_button
 from src.launchers.launcher_manager_attrs import forward_manager_attribute
+from src.launchers._launcher_navigation_ui import (
+    LauncherNavigationUIMixin,
+    _build_menu_bar_close_widget as _navigation_menu_bar_close_widget,
+)
+from src.launchers._launcher_top_bar_ui import (
+    ClickableLabel as ClickableLabel,
+    HelpButtonHoverFilter as HelpButtonHoverFilter,
+    LauncherTopBarUIMixin,
+    RuntimeButton as RuntimeButton,
+    _build_zoom_accessible_description as _top_bar_zoom_description,
+)
 
 from src.shared.python.logging_pkg.logging_config import get_logger
 from src.shared.python.theme.style_constants import Styles
-from src.shared.python.theme.typography import Weights, get_display_font
-from src.shared.python.ui.auto_complete import AutoCompleteLineEdit
-from src.shared.python.ui.completion_vocab import build_vocabulary
 
 if TYPE_CHECKING:
     pass
@@ -61,94 +62,18 @@ if TYPE_CHECKING:
 logger = get_logger(__name__)
 
 
-def _resolve_theme_color(c_obj: Any, ns_key: str, dict_key: str, default: str) -> str:
-    """Helper to resolve colors from namespace/dict theme tokens."""
-    if isinstance(c_obj, dict):
-        return c_obj.get(dict_key, default)
-    return getattr(c_obj, ns_key, getattr(c_obj, dict_key, default))
+def _build_zoom_accessible_description() -> str:
+    """Describe zoom bounds through the historical facade helper."""
+    return _top_bar_zoom_description(TILE_SCALE_MIN, TILE_SCALE_MAX)
 
 
-class ClickableLabel(QLabel):
-    """QLabel subclass that emits a clicked signal on mouse click."""
-
-    clicked = pyqtSignal()
-
-    def mousePressEvent(self, event: QMouseEvent) -> None:
-        self.clicked.emit()
-        super().mousePressEvent(event)
-
-    def enterEvent(self, event: QEnterEvent) -> None:
-        if self.cursor().shape() == Qt.CursorShape.PointingHandCursor:
-            font = self.font()
-            font.setUnderline(True)
-            self.setFont(font)
-        super().enterEvent(event)
-
-    def leaveEvent(self, event: QEvent) -> None:
-        font = self.font()
-        font.setUnderline(False)
-        self.setFont(font)
-        super().leaveEvent(event)
-
-
-class RuntimeButton(QToolButton):
-    """QToolButton subclass designed to mimic model card tiles.
-
-    Supports custom drop shadow, theme-matching styling, and deferred hover-hide behavior
-    for its associated help button.
-    """
-
-    def __init__(
-        self, parent: QWidget | None = None, help_button: QWidget | None = None
-    ) -> None:
-        super().__init__(parent)
-        self.help_button = help_button
-        self.setCursor(Qt.CursorShape.PointingHandCursor)
-        self.setObjectName("RuntimeButton")
-        self.setToolButtonStyle(Qt.ToolButtonStyle.ToolButtonTextOnly)
-
-        # Reduced/extremely subtle shadow
-        self.shadow = QGraphicsDropShadowEffect(self)
-        self.shadow.setBlurRadius(2)
-        self.shadow.setOffset(0, 1)
-        self.shadow.setColor(QColor(0, 0, 0, 15))
-        self.setGraphicsEffect(self.shadow)
-
-        # Deferred hide timer to allow moving mouse to the help button
-        self._hide_timer = QTimer(self)
-        self._hide_timer.setSingleShot(True)
-        self._hide_timer.setInterval(300)  # 300ms delay
-        self._hide_timer.timeout.connect(self._do_hide)
-
-    def enterEvent(self, event: QEnterEvent) -> None:
-        self._hide_timer.stop()
-        if self.help_button:
-            self.help_button.show()
-            self.help_button.raise_()
-        super().enterEvent(event)
-
-    def leaveEvent(self, event: QEvent) -> None:
-        self._hide_timer.start()
-        super().leaveEvent(event)
-
-    def _do_hide(self) -> None:
-        if self.help_button and not self.help_button.underMouse():
-            self.help_button.hide()
-
-
-class HelpButtonHoverFilter(QObject):
-    """Event filter to prevent hiding the help button when hovered directly."""
-
-    def __init__(self, runtime_button: RuntimeButton) -> None:
-        super().__init__()
-        self.runtime_button = runtime_button
-
-    def eventFilter(self, obj: QObject, event: QEvent) -> bool:
-        if event.type() == QEvent.Type.Enter:
-            self.runtime_button._hide_timer.stop()
-        elif event.type() == QEvent.Type.Leave:
-            self.runtime_button._hide_timer.start()
-        return super().eventFilter(obj, event)
+def _build_menu_bar_close_widget(parent: QWidget, close_callback: Any) -> QWidget:
+    """Build the historical menu close control with a patchable factory."""
+    return _navigation_menu_bar_close_widget(
+        parent,
+        close_callback,
+        button_factory=create_window_control_button,
+    )
 
 
 class ProcessOutputRelay(QObject):
@@ -174,36 +99,6 @@ class ProcessOutputRelay(QObject):
     def _deliver(self, engine_name: str, line: str) -> None:
         """Forward one line to the console sink (GUI thread only)."""
         self._sink(engine_name, line)
-
-
-def _build_zoom_accessible_description() -> str:
-    """Describe the zoom slider using the configured tile-scale bounds."""
-    minimum_pct = int(round(TILE_SCALE_MIN * 100))
-    maximum_pct = int(round(TILE_SCALE_MAX * 100))
-    return (
-        f"Adjust tile size from {minimum_pct}% to {maximum_pct}%. "
-        "Use arrow keys or drag to adjust."
-    )
-
-
-def _build_menu_bar_close_widget(parent: QWidget, close_callback: Any) -> QWidget:
-    """Create the top-row close control for the launcher menu bar."""
-    container = QWidget(parent)
-    layout = QHBoxLayout(container)
-    layout.setContentsMargins(0, 1, 6, 1)
-    layout.setSpacing(0)
-
-    close_button = create_window_control_button(
-        "close",
-        "X",
-        tooltip="Close the launcher",
-        accessible_name="Close launcher window",
-        object_name="menu-bar-close-button",
-        parent=container,
-    )
-    close_button.clicked.connect(lambda _checked=False: close_callback())
-    layout.addWidget(close_button)
-    return container
 
 
 from typing import Protocol
@@ -287,7 +182,7 @@ class ResizingScrollArea(QScrollArea):
                     launcher._rebuild_grid()
 
 
-class UISetupManager:
+class UISetupManager(LauncherNavigationUIMixin, LauncherTopBarUIMixin):
     def __init__(self, launcher):
         self.launcher = launcher
 
@@ -305,61 +200,24 @@ class UISetupManager:
             return
         handler(state)
 
+    def wheelEvent(self, event: Any) -> None:  # noqa: N802
+        """Ctrl+scroll wheel adjusts the zoom slider."""
+        from PyQt6.QtCore import Qt as _Qt
+
+        if event.modifiers() & _Qt.KeyboardModifier.ControlModifier:
+            delta = event.angleDelta().y()
+            if delta != 0:
+                steps = 5 if delta > 0 else -5
+                self._nudge_zoom(steps)
+                event.accept()
+                return
+        super().wheelEvent(event)  # type: ignore[misc]
+
     """Mixin for UpstreamDriftLauncher UI initialization.
 
     Provides methods for building the menu bar, top bar, grid area,
     bottom bar, search shortcuts, process console, context help, and AI panel.
     """
-
-    def _build_sidebar_button(
-        self,
-        label: str,
-        icon_name: str,
-        *,
-        checkable: bool = False,
-    ) -> QToolButton:
-        """Create an icon-first sidebar control with accessible labeling.
-
-        Provides both icon and visible text label for accessibility.
-        """
-        button = QToolButton()
-        button.setText(label)
-        button.setToolTip(label)
-        button.setAccessibleName(label)
-        button.setAccessibleDescription(f"Navigate to {label} section")
-        button.setCursor(Qt.CursorShape.PointingHandCursor)
-        # Show both icon and text for better accessibility
-        button.setToolButtonStyle(Qt.ToolButtonStyle.ToolButtonTextUnderIcon)
-
-        # Sidebar buttons must always render a visible icon (#5624).
-        # If the requested glyph is missing from SVG_REGISTRY, fall back
-        # to a registered glyph so the icon is never null — null icons
-        # produce the text-only sidebar regression visible in #5624.
-        try:
-            from src.shared.python.theme.icon_utils import (
-                SVG_REGISTRY,
-                IconColorizer,
-            )
-
-            resolved_name = icon_name if icon_name in SVG_REGISTRY else "settings"
-            button.setIcon(IconColorizer.get_icon(resolved_name, "#d4d4d4"))
-        except (ImportError, ValueError):
-            # Last-resort fallback: any QIcon (even with a tinted blank
-            # pixmap) is preferable to a null icon for hit-testing and
-            # screen-reader semantics.
-            from PyQt6.QtGui import QIcon, QPixmap
-
-            pixmap = QPixmap(22, 22)
-            pixmap.fill(Qt.GlobalColor.transparent)
-            button.setIcon(QIcon(pixmap))
-
-        button.setIconSize(QSize(22, 22))
-        button.setCheckable(checkable)
-        button.setAutoRaise(True)
-        # Set focus policy for keyboard navigation
-        button.setFocusPolicy(Qt.FocusPolicy.StrongFocus)
-        button.setSizePolicy(QSizePolicy.Policy.Expanding, QSizePolicy.Policy.Fixed)
-        return button
 
     def init_ui(self) -> None:
         """Initialize the user interface.
@@ -756,1152 +614,15 @@ class UISetupManager:
         # Initialize Overlay
         self._init_overlay()
 
-    def _setup_global_sidebar(self) -> QWidget:
-        """Create the global sidebar navigation."""
-        sidebar = QWidget()
-        sidebar.setMinimumWidth(Styles.SIDEBAR_MIN_WIDTH)
-        sidebar.setSizePolicy(
-            QSizePolicy.Policy.Preferred, QSizePolicy.Policy.Expanding
-        )
-
-        try:
-            from src.shared.python.theme import get_current_colors
-
-            colors = get_current_colors()
-        except ImportError:
-            from src.shared.python.theme import DARK_THEME
-
-            colors = DARK_THEME
-
-        sidebar.setStyleSheet(f"""
-            QWidget {{
-                background-color: {colors.bg_elevated};
-                border-right: 1px solid {colors.border_default};
-            }}
-            QToolButton {{
-                background-color: transparent;
-                border: 1px solid transparent;
-                border-radius: 8px;
-                color: {colors.text_secondary};
-                padding: 12px 0;
-            }}
-            QToolButton:hover {{
-                background-color: {colors.bg_highlight};
-                color: {colors.text_primary};
-                border: 1px solid {colors.border_default};
-            }}
-            QToolButton:checked {{
-                background-color: {colors.primary};
-                color: {colors.bg};
-                border: 1px solid {colors.primary};
-            }}
-        """)
-
-        layout = QVBoxLayout(sidebar)
-        layout.setContentsMargins(8, 20, 8, 20)
-        layout.setSpacing(15)
-
-        btn_home = self._build_sidebar_button(
-            "Home",
-            "home",
-            checkable=True,
-        )
-        btn_home.setChecked(True)
-
-        btn_engines = self._build_sidebar_button(
-            "Engines",
-            "computer",
-            checkable=True,
-        )
-
-        btn_biomechanics = self._build_sidebar_button(
-            "Biomechanics",
-            "accessibility",
-            checkable=True,
-        )
-        btn_biomechanics.setAccessibleDescription(
-            "Filter tiles to show biomechanics and motion analysis tools"
-        )
-
-        btn_simulation = self._build_sidebar_button(
-            "Simulation",
-            "sports_golf",
-            checkable=True,
-        )
-
-        btn_motion_match = self._build_sidebar_button(
-            "Motion Match",
-            "directions_run",
-            checkable=True,
-        )
-        btn_motion_match.setAccessibleDescription(
-            "Filter tiles to show motion matching tools"
-        )
-
-        btn_mocap = self._build_sidebar_button(
-            "MoCap",
-            "videocam",
-            checkable=True,
-        )
-        btn_mocap.setAccessibleDescription("Filter tiles to show motion capture tools")
-
-        btn_tools = self._build_sidebar_button(
-            "Tools",
-            "build",
-            checkable=True,
-        )
-
-        btn_documentation = self._build_sidebar_button(
-            "Documentation",
-            "book",
-            checkable=True,
-        )
-        btn_documentation.setAccessibleDescription(
-            "Filter tiles to show documentation and library tools"
-        )
-
-        btn_favorites = self._build_sidebar_button(
-            "Favorites",
-            "star",
-            checkable=True,
-        )
-        btn_favorites.setAccessibleDescription("Filter tiles to show your favorites")
-
-        btn_history = self._build_sidebar_button(
-            "History",
-            "history",
-            checkable=True,
-        )
-        btn_history.setAccessibleDescription(
-            "Filter tiles to show recently and frequently used tools"
-        )
-
-        btn_settings = self._build_sidebar_button(
-            "Settings",
-            "settings",
-            checkable=False,
-        )
-        if True:
-            btn_settings.clicked.connect(self._show_preferences)
-
-        btn_console = self._build_sidebar_button(
-            "Console",
-            "terminal",
-            checkable=False,
-        )
-        btn_console.setAccessibleDescription(
-            "Show or hide the process output console window"
-        )
-        btn_console.clicked.connect(self.toggle_process_console)
-        self.btn_console = btn_console
-
-        # Setup mutually exclusive active-state routing for navigation
-        self.sidebar_group = QButtonGroup(self.launcher)
-        self.sidebar_group.addButton(btn_home, 0)
-        self.sidebar_group.addButton(btn_engines, 1)
-        self.sidebar_group.addButton(btn_biomechanics, 2)
-        self.sidebar_group.addButton(btn_simulation, 3)
-        self.sidebar_group.addButton(btn_motion_match, 4)
-        self.sidebar_group.addButton(btn_mocap, 5)
-        self.sidebar_group.addButton(btn_tools, 6)
-        self.sidebar_group.addButton(btn_documentation, 7)
-        self.sidebar_group.addButton(btn_favorites, 8)
-        self.sidebar_group.addButton(btn_history, 9)
-        self.sidebar_group.idClicked.connect(self._on_sidebar_routed)
-
-        btn_library = self._build_sidebar_button(
-            "Library",
-            "book",
-            checkable=True,
-        )
-        btn_library.setAccessibleDescription(
-            "Filter tiles to show documentation and library tools"
-        )
-        self.sidebar_group.addButton(btn_library, 10)
-        self.btn_library_sidebar = btn_library
-        self.btn_training_sidebar = None
-
-        # Space navigation buttons evenly to fill available height
-        layout.addWidget(btn_home)
-        layout.addStretch(1)
-        layout.addWidget(btn_engines)
-        layout.addStretch(1)
-        layout.addWidget(btn_biomechanics)
-        layout.addStretch(1)
-        layout.addWidget(btn_simulation)
-        layout.addStretch(1)
-        layout.addWidget(btn_motion_match)
-        layout.addStretch(1)
-        layout.addWidget(btn_mocap)
-        layout.addStretch(1)
-        layout.addWidget(btn_tools)
-        layout.addStretch(1)
-        layout.addWidget(btn_documentation)
-        layout.addStretch(1)
-        layout.addWidget(btn_library)
-        layout.addStretch(1)
-        layout.addWidget(btn_favorites)
-        layout.addStretch(1)
-        layout.addWidget(btn_history)
-        layout.addStretch(3)  # larger gap before bottom group
-        layout.addWidget(btn_console)
-        layout.addWidget(btn_settings)
-
-        # Set explicit focus order for keyboard navigation
-        sidebar.setFocusProxy(btn_home)
-        QWidget.setTabOrder(btn_home, btn_engines)
-        QWidget.setTabOrder(btn_engines, btn_biomechanics)
-        QWidget.setTabOrder(btn_biomechanics, btn_simulation)
-        QWidget.setTabOrder(btn_simulation, btn_motion_match)
-        QWidget.setTabOrder(btn_motion_match, btn_mocap)
-        QWidget.setTabOrder(btn_mocap, btn_tools)
-        QWidget.setTabOrder(btn_tools, btn_documentation)
-        QWidget.setTabOrder(btn_documentation, btn_library)
-        QWidget.setTabOrder(btn_library, btn_favorites)
-        QWidget.setTabOrder(btn_favorites, btn_history)
-        QWidget.setTabOrder(btn_history, btn_console)
-        QWidget.setTabOrder(btn_console, btn_settings)
-
-        scroll_area = QScrollArea()
-        scroll_area.setWidget(sidebar)
-        scroll_area.setWidgetResizable(True)
-        scroll_area.setHorizontalScrollBarPolicy(Qt.ScrollBarPolicy.ScrollBarAlwaysOff)
-        scroll_area.setMinimumWidth(Styles.SIDEBAR_MIN_WIDTH)
-        scroll_area.setFrameShape(QFrame.Shape.NoFrame)
-        scroll_area.setStyleSheet(f"""
-            QScrollArea {{
-                background-color: {colors.bg_elevated};
-                border-right: 1px solid {colors.border_default};
-            }}
-        """)
-
-        return scroll_area
-
-    def _on_sidebar_routed(self, button_id: int) -> None:
-        """Route sidebar navigation to filter the grid layout.
-
-        Button IDs
-        ----------
-        0  All (Home)
-        1  Engines
-        2  Biomechanics
-        3  Simulation
-        4  Motion Match
-        5  MoCap
-        6  Tools
-        7  Documentation
-        8  Favorites
-        9  History
-        10 Library
-        """
-        _CATEGORY_MAP: dict[int, str] = {
-            0: "All",
-            1: "Engines",
-            2: "Biomechanics",
-            3: "Simulation",
-            4: "Tools",
-            5: "Tools",
-            6: "Tools",
-            7: "Documentation",
-            8: "Favorites",
-            9: "History",
-            10: "Documentation",
-        }
-        self.layout_manager.current_category_filter = _CATEGORY_MAP.get(
-            button_id, "All"
-        )
-
-        self._rebuild_grid()
-
-        if getattr(self, "workspace_tabs", None):
-            self.workspace_tabs.setCurrentIndex(0)
-
-    def _build_menu_bar_widget(self) -> QMenuBar:
-        """Build a populated ``QMenuBar`` for the frameless launcher (#5624).
-
-        Returns a standalone ``QMenuBar`` (parent ``self``) that the
-        caller adds to the central widget's outer ``QVBoxLayout`` so it
-        sits **below** the custom title bar.
-
-        Postcondition: the returned widget is non-null, parented to
-        ``self``, and populated with the File/View/Tools/Help menus.
-
-        We deliberately do not call ``QMainWindow.setMenuBar`` — that
-        reserves the native top strip above the central widget, which
-        on a frameless window draws above the custom title bar.
-        """
-        menubar = QMenuBar(self.launcher)
-        # Postcondition (DbC): a non-null QMenuBar is returned.
-        assert menubar is not None, (
-            "QMenuBar construction returned None — should be impossible"
-        )
-
-        self._setup_file_menu(menubar)
-        self._setup_view_menu(menubar)
-        self._setup_tools_menu(menubar)
-        self._setup_help_menu(menubar)
-        return menubar
-
-    def _setup_menu_bar(self) -> None:
-        """Set up the application menu bar.
-
-        .. deprecated:: #5624
-            Kept for backwards compatibility with any caller still
-            reaching for the legacy hook.  Prefer
-            ``_build_menu_bar_widget`` + adding the result to the
-            central widget's outer layout.
-        """
-        menubar = self.menuBar()
-
-        self._setup_file_menu(menubar)
-        self._setup_view_menu(menubar)
-        self._setup_tools_menu(menubar)
-        self._setup_help_menu(menubar)
-        menubar.setCornerWidget(
-            _build_menu_bar_close_widget(self, self.close),
-            Qt.Corner.TopRightCorner,
-        )
-
-    def _setup_file_menu(self, menubar: Any) -> None:
-        if menubar is None:
-            raise ValueError("menubar must be provided")
-        file_menu = menubar.addMenu("&File")
-
-        action_preferences = QAction("&Preferences...", self.launcher)
-        action_preferences.setShortcut("Ctrl+,")
-        action_preferences.setToolTip("Edit application preferences and theme")
-        action_preferences.setStatusTip("Opens preferences")
-        action_preferences.triggered.connect(self._show_preferences)
-        file_menu.addAction(action_preferences)
-
-        file_menu.addSeparator()
-
-        action_exit = QAction("E&xit", self.launcher)
-        action_exit.setShortcut("Ctrl+Q")
-        action_exit.setToolTip("Close the launcher")
-        action_exit.setStatusTip("Quits the application")
-        action_exit.triggered.connect(self.close)
-        file_menu.addAction(action_exit)
-
-    def _setup_view_menu(self, menubar: Any) -> None:
-        if menubar is None:
-            raise ValueError("menubar must be provided")
-        view_menu = menubar.addMenu("&View")
-
-        from PyQt6.QtGui import QActionGroup
-
-        self._viewmode_action_group = QActionGroup(self.launcher)
-        self._viewmode_action_group.setExclusive(True)
-        self._viewmode_actions: dict[ViewMode, QAction] = {}
-        for label, mode, shortcut in (
-            ("Tile &Small", ViewMode.SMALL, "Ctrl+1"),
-            ("Tile &Medium", ViewMode.MEDIUM, "Ctrl+2"),
-            ("Tile &Large", ViewMode.LARGE, "Ctrl+3"),
-            ("List &Small", ViewMode.LIST_SMALL, "Ctrl+4"),
-            ("List &Large", ViewMode.LIST_LARGE, "Ctrl+5"),
-        ):
-            act = QAction(label, self.launcher)
-            act.setCheckable(True)
-            act.setShortcut(shortcut)
-            act.setToolTip(f"Switch tile layout to {label.replace('&', '')} mode")
-            act.setStatusTip(act.toolTip())
-            act.triggered.connect(
-                lambda _checked=False, m=mode: self._set_view_mode_from_menu(m)
-            )
-            self._viewmode_action_group.addAction(act)
-            view_menu.addAction(act)
-            self._viewmode_actions[mode] = act
-        # Default checkmark on LIST_LARGE.
-        self._viewmode_actions[ViewMode.LIST_LARGE].setChecked(True)
-
-        view_menu.addSeparator()
-
-        action_layout_mode = QAction("&Edit Layout Mode", self.launcher)
-        action_layout_mode.setCheckable(True)
-        action_layout_mode.setToolTip("Toggle drag-and-drop reordering of model tiles")
-        action_layout_mode.setStatusTip("Toggles layout-edit mode")
-        action_layout_mode.triggered.connect(self._toggle_layout_mode_from_menu)
-        view_menu.addAction(action_layout_mode)
-        self._action_layout_mode = action_layout_mode
-
-        action_customize_tiles = QAction("&Select Visible Tiles...", self.launcher)
-        action_customize_tiles.setToolTip(
-            "Select which tiles are visible in the layout"
-        )
-        action_customize_tiles.setStatusTip("Select visible tiles")
-        action_customize_tiles.triggered.connect(self.open_layout_manager)
-        view_menu.addAction(action_customize_tiles)
-        self.action_customize_tiles = action_customize_tiles
-
-        view_menu.addSeparator()
-
-        action_context_help = QAction("Context &Help Panel", self.launcher)
-        action_context_help.setCheckable(True)
-        action_context_help.setToolTip("Show or hide the side help panel")
-        action_context_help.setStatusTip("Toggles context-help panel")
-        action_context_help.triggered.connect(self._toggle_context_help)
-        view_menu.addAction(action_context_help)
-        self._action_context_help = action_context_help
-
-        action_console = QAction("&Process Output Console", self.launcher)
-        action_console.setCheckable(True)
-        action_console.setChecked(False)
-        action_console.setShortcut("Ctrl+`")
-        action_console.setToolTip("Show or hide the launched-process output console")
-        action_console.setStatusTip("Toggles process-output console")
-        action_console.triggered.connect(lambda checked: self.toggle_process_console())
-        view_menu.addAction(action_console)
-        self._action_console = action_console
-
-        view_menu.addSeparator()
-
-        action_popout_tab = QAction("&Undock Active Tab", self.launcher)
-        action_popout_tab.setShortcut("Ctrl+D")
-        action_popout_tab.setToolTip("Detach the active tab into a floating window")
-        action_popout_tab.setStatusTip("Undock active tab")
-        action_popout_tab.triggered.connect(self._popout_active_tab)
-        view_menu.addAction(action_popout_tab)
-        self.action_popout_tab = action_popout_tab
-
-        action_redock_tabs = QAction("&Redock All Tabs", self.launcher)
-        action_redock_tabs.setShortcut("Ctrl+Shift+D")
-        action_redock_tabs.setToolTip(
-            "Redock all floating windows back to workspace tabs"
-        )
-        action_redock_tabs.setStatusTip("Redock all tabs")
-        action_redock_tabs.triggered.connect(self._redock_all_tabs)
-        view_menu.addAction(action_redock_tabs)
-        self.action_redock_tabs = action_redock_tabs
-
-        background_menu = view_menu.addMenu("&Background Tabs")
-        self._background_tabs_menu = background_menu
-        background_menu.aboutToShow.connect(self._refresh_background_tabs_menu)
-
-        view_menu.addSeparator()
-        theme_menu = view_menu.addMenu("&Theme")
-        self._setup_theme_menu(theme_menu)
-
-    def _refresh_background_tabs_menu(self) -> None:
-        """Populate the Background Tabs menu with restore actions (#6013).
-
-        Lists tool tabs that were closed but kept running hidden, so the
-        user can bring them back. Rebuilt each time the menu opens.
-        """
-        menu = getattr(self, "_background_tabs_menu", None)
-        if menu is None:
-            return
-        menu.clear()
-
-        tabs = getattr(self, "workspace_tabs", None)
-        titles = tabs.list_background_tabs() if tabs is not None else []
-
-        if not titles:
-            empty = QAction("(no background tabs)", self.launcher)
-            empty.setEnabled(False)
-            menu.addAction(empty)
-            return
-
-        for title in titles:
-            act = QAction(f"Restore: {title}", self.launcher)
-            act.setToolTip(f"Bring the hidden '{title}' tab back into view")
-            act.triggered.connect(
-                lambda _checked=False, t=title: tabs.restore_background_tab(t)
-            )
-            menu.addAction(act)
-
-        menu.addSeparator()
-        restore_all = QAction("Restore All", self.launcher)
-        restore_all.triggered.connect(tabs.restore_all_background_tabs)
-        menu.addAction(restore_all)
-
-    def _popout_active_tab(self) -> None:
-        """Undock the currently active workspace tab into a floating window."""
-        if not getattr(self, "workspace_tabs", None):
-            return
-        idx = self.workspace_tabs.currentIndex()
-        if idx < 0:
-            return
-        tab_text = self.workspace_tabs.tabText(idx)
-        if tab_text == "Home":
-            from PyQt6.QtWidgets import QMessageBox
-
-            QMessageBox.information(
-                self.launcher,
-                "Cannot Undock",
-                "The 'Home' tab is a core view and cannot be undocked.",
-            )
-            return
-        self.workspace_tabs.detach_tab_from_menu(idx)
-
-    def _redock_all_tabs(self) -> None:
-        """Redock all detached workspace tabs."""
-        if getattr(self, "workspace_tabs", None):
-            self.workspace_tabs.redock_all_tabs()
-
-    def _setup_tools_menu(self, menubar: Any) -> None:
-        if menubar is None:
-            raise ValueError("menubar must be provided")
-        tools_menu = menubar.addMenu("&Tools")
-
-        action_env = QAction("&Environment Manager...", self.launcher)
-        action_env.setToolTip("Inspect Python environments and engine availability")
-        action_env.setStatusTip("Opens environment manager")
-        action_env.triggered.connect(lambda: self._open_settings(tab=1))
-        tools_menu.addAction(action_env)
-
-        action_diag = QAction("&Diagnostics...", self.launcher)
-        action_diag.setToolTip("Run a diagnostic sweep of the install")
-        action_diag.setStatusTip("Opens diagnostics")
-        action_diag.triggered.connect(lambda: self._open_settings(tab=2))
-        tools_menu.addAction(action_diag)
-
-    def _setup_help_menu(self, menubar: Any) -> None:
-        if menubar is None:
-            raise ValueError("menubar must be provided")
-        help_menu = menubar.addMenu("&Help")
-
-        # User Manual + topic-help actions live under the legacy in-app help
-        # dialog.  The new structured Help submenu (User Guide, Loaders,
-        # Keyboard Shortcuts, Report a Bug, About) is appended after it via
-        # build_help_menu_into so the entries gain consistent tooltips and
-        # status tips.
-        from src.launchers.about_dialog import (
-            open_issues_page,
-            open_motion_match_loaders_doc,
-            open_user_guide,
-            show_about_dialog,
-        )
-        from src.launchers.help_menu import show_keyboard_shortcuts_modal
-
-        action_manual = QAction("&User Manual", self.launcher)
-        action_manual.setShortcut("F1")
-        action_manual.setToolTip("Open the in-app help dialog")
-        action_manual.setStatusTip("Opens user manual")
-        action_manual.triggered.connect(lambda: self._show_help_dialog())
-        help_menu.addAction(action_manual)
-
-        action_context_docs = QAction("Context &Documentation", self.launcher)
-        action_context_docs.setToolTip("Open context-aware documentation")
-        action_context_docs.setStatusTip("Opens Context Help")
-        if True:
-            action_context_docs.triggered.connect(self._toggle_context_help)
-        help_menu.addAction(action_context_docs)
-
-        action_user_guide = QAction("User &Guide (online)", self.launcher)
-        action_user_guide.setToolTip(
-            "Open the bundled user guide in the system browser"
-        )
-        action_user_guide.setStatusTip("Opens user guide")
-        action_user_guide.triggered.connect(lambda: open_user_guide())
-        help_menu.addAction(action_user_guide)
-
-        action_loaders = QAction("&Motion-Match Loaders", self.launcher)
-        action_loaders.setToolTip("Reference for loading motion-target files")
-        action_loaders.setStatusTip("Opens motion-match loader reference")
-        action_loaders.triggered.connect(lambda: open_motion_match_loaders_doc())
-        help_menu.addAction(action_loaders)
-
-        action_project_map = QAction("&Project Map", self.launcher)
-        action_project_map.setToolTip("Open the project-map document")
-        action_project_map.setStatusTip("Opens project map")
-        action_project_map.triggered.connect(self._open_project_map)
-        help_menu.addAction(action_project_map)
-
-        help_menu.addSeparator()
-
-        topic_tips: dict[str, tuple[str, str]] = {
-            "engine_selection": (
-                "How to choose between the bundled physics engines",
-                "Opens engine-selection help",
-            ),
-            "simulation_controls": (
-                "Reference for simulation playback and stepping controls",
-                "Opens simulation-controls help",
-            ),
-            "motion_capture": (
-                "Loading and aligning motion-capture targets",
-                "Opens motion-capture help",
-            ),
-            "visualization": (
-                "Camera, traces, and scene-element controls",
-                "Opens visualization help",
-            ),
-            "analysis_tools": (
-                "Built-in analysis and post-processing tools",
-                "Opens analysis-tools help",
-            ),
-        }
-        for label, topic in [
-            ("Engine &Selection Guide", "engine_selection"),
-            ("Simulation &Controls", "simulation_controls"),
-            ("Motion &Capture", "motion_capture"),
-            ("&Visualization", "visualization"),
-            ("&Analysis Tools", "analysis_tools"),
-        ]:
-            action = QAction(label, self.launcher)
-            tip, status = topic_tips[topic]
-            action.setToolTip(tip)
-            action.setStatusTip(status)
-            action.triggered.connect(lambda checked, t=topic: self._show_help_dialog(t))
-            help_menu.addAction(action)
-
-        help_menu.addSeparator()
-
-        action_shortcuts = QAction("&Keyboard Shortcuts...", self.launcher)
-        action_shortcuts.setShortcut("Ctrl+?")
-        action_shortcuts.setToolTip("Show every registered keyboard shortcut")
-        action_shortcuts.setStatusTip("Opens keyboard-shortcuts table")
-
-        # Prefer the structured modal that scrapes live actions; fall
-        # back to the legacy overlay if the modal raises.
-        def _open_shortcuts() -> None:
-            try:
-                show_keyboard_shortcuts_modal(self.launcher)
-            except Exception:  # noqa: BLE001
-                self._show_shortcuts_overlay()
-
-        action_shortcuts.triggered.connect(_open_shortcuts)
-        help_menu.addAction(action_shortcuts)
-
-        action_report_bug = QAction("&Report a Bug...", self.launcher)
-        action_report_bug.setToolTip("Open the public issue tracker in your browser")
-        action_report_bug.setStatusTip("Opens issue tracker")
-        action_report_bug.triggered.connect(lambda: open_issues_page())
-        help_menu.addAction(action_report_bug)
-
-        help_menu.addSeparator()
-
-        action_about = QAction("&About UpstreamDrift", self.launcher)
-        action_about.setToolTip("Show version and runtime information")
-        action_about.setStatusTip("Opens About dialog")
-
-        # Prefer the new dialog with live versions; fall back to legacy
-        # if anything fails (keeps menu working in trimmed environments).
-        def _open_about() -> None:
-            try:
-                show_about_dialog(self.launcher)
-            except Exception:  # noqa: BLE001
-                self._show_about_dialog()
-
-        action_about.triggered.connect(_open_about)
-        help_menu.addAction(action_about)
-
-    def _setup_top_bar_status_and_search(self, top_bar: QHBoxLayout) -> None:
-        """Add status indicator, execution mode label, and search bar to top bar."""
-        # Status Indicator
-        if top_bar is None:
-            raise ValueError("top_bar must be provided")
-        self.lbl_status = ClickableLabel("Checking Docker...")
-        self.lbl_status.setProperty("status", "inactive-bold")
-        _style = self.lbl_status.style()
-
-        if _style:
-            _style.polish(self.lbl_status)
-        self.lbl_status.clicked.connect(self._on_status_clicked)
-        top_bar.addWidget(self.lbl_status)
-
-        # Engine-runtime indicator. Shows where physics engines run:
-        # Native Windows (host Python), Docker container, or WSL2.
-        from src.launchers.runtime_mode_help import make_runtime_mode_help_button
-
-        self.btn_runtime_help = make_runtime_mode_help_button(self.launcher)
-        self.btn_runtime_help.hide()  # Hidden by default, shown on hover
-
-        try:
-            import src.shared.python.theme as _theme
-
-            c = _theme.get_current_colors()  # type: ignore[attr-defined]
-        except (ImportError, AttributeError):
-            from src.shared.python.theme import DARK_THEME as c  # type: ignore[assignment,no-redef]
-
-        bg = _resolve_theme_color(c, "surface_hover", "group_bg", "#2d2d2d")
-        border = _resolve_theme_color(c, "border_light", "border", "#444444")
-        bg_hover = _resolve_theme_color(c, "surface_active", "input_bg", "#3a3a3a")
-        border_hover = _resolve_theme_color(c, "border_strong", "focus", "#666666")
-        text_color = _resolve_theme_color(c, "text_primary", "text", "#ffffff")
-
-        # Engine-runtime indicator is now a styled RuntimeButton with shadow, border, and hover help triggers
-        self.lbl_execution_mode = RuntimeButton(help_button=self.btn_runtime_help)
-        self.lbl_execution_mode.setText("Runtime: Windows")
-        self.lbl_execution_mode.setProperty("exec_mode", "warning")
-        _style = self.lbl_execution_mode.style()
-        if _style:
-            _style.polish(self.lbl_execution_mode)
-        self.lbl_execution_mode.setToolTip(
-            "Click to change the engine runtime (Windows, Docker, or "
-            "WSL2 Ubuntu) in Settings → Configuration. Hover to explain."
-        )
-        self.lbl_execution_mode.setStyleSheet(f"""
-            QToolButton#RuntimeButton {{
-                background-color: {bg};
-                border: 1px solid {border};
-                border-radius: 8px;
-                padding: 4px 12px;
-                color: {text_color};
-                font-weight: bold;
-                margin-left: 10px;
-            }}
-            QToolButton#RuntimeButton:hover {{
-                background-color: {bg_hover};
-                border: 1px solid {border_hover};
-            }}
-        """)
-        # tab=1 → Configuration tab, which hosts the Engine Runtime group.
-        self.lbl_execution_mode.clicked.connect(
-            lambda: self.launcher._open_settings(tab=1)
-        )
-        top_bar.addWidget(self.lbl_execution_mode)
-
-        # Wire event filter to help button to prevent hiding it when hovered directly
-        self._help_button_filter = HelpButtonHoverFilter(self.lbl_execution_mode)
-        self.btn_runtime_help.installEventFilter(self._help_button_filter)
-        top_bar.addWidget(self.btn_runtime_help)
-
-        top_bar.addStretch()
-
-        # Search Bar
-        self.search_input = AutoCompleteLineEdit(words=build_vocabulary())
-        self.search_input.setPlaceholderText("Search models... (Esc to clear)")
-        try:
-            from src.shared.python.theme.responsive import (
-                set_text_minimum_width,
-                TextWidthSpec,
-            )
-
-            set_text_minimum_width(
-                self.search_input,
-                TextWidthSpec(minimum_px=250),
-            )
-        except ImportError:
-            self.search_input.setFixedWidth(250)
-        self.search_input.setToolTip("Filter models by name or description (Ctrl+F)")
-        self.search_input.setAccessibleName("Search models")
-        self.search_input.setClearButtonEnabled(True)
-        self.search_input.textChanged.connect(self.update_search_filter)
-        top_bar.addWidget(self.search_input)
-
-        # Clear Filters Button
-        self.btn_clear_filters = QPushButton("Clear Filters")
-        self.btn_clear_filters.setCursor(Qt.CursorShape.PointingHandCursor)
-        self.btn_clear_filters.setToolTip(
-            "Reset all category and search filters to show all tiles"
-        )
-        self.btn_clear_filters.setStyleSheet("""
-            QPushButton {
-                background-color: rgba(255, 255, 255, 0.05);
-                border: 1px solid rgba(255, 255, 255, 0.15);
-                border-radius: 4px;
-                padding: 4px 10px;
-                color: #e0e0e0;
-                font-size: 12px;
-            }
-            QPushButton:hover {
-                background-color: rgba(255, 255, 255, 0.12);
-                border-color: rgba(255, 255, 255, 0.25);
-                color: #ffffff;
-            }
-        """)
-        self.btn_clear_filters.clicked.connect(self._clear_all_filters)
-        self.btn_clear_filters.hide()
-        top_bar.addWidget(self.btn_clear_filters)
-
-        # Launch button — sits in the top bar next to the search input so the
-        # primary action is always reachable without scrolling and the grid
-        # area below is freed for tiles (no bottom-bar overlap).
-        self._ensure_launch_button()
         # top_bar.addWidget(self.btn_launch)  # Removed launch button per user request
-
-    def _on_status_clicked(self) -> None:
-        """Handle clicking the status label. If in error/dependency error, opens environment manager.
-        If there are running processes, routes to the Processes tab in Settings.
-        """
-        status_text = self.lbl_status.text()
-        if "Dependency Error" in status_text or "Error" in status_text:
-            self.launcher.open_environment_manager()
-            return
-
-        # Check active processes
-        running = []
-        with self.process_manager._process_lock:
-            for name, proc in list(self.running_processes.items()):
-                if proc.poll() is None:
-                    running.append(name)
-
-        if running:
-            self._open_settings(tab=8)
-
-    def _ensure_launch_button(self) -> None:
-        """Create ``self.btn_launch`` if it doesn't exist yet (idempotent).
-
-        Used by both ``_setup_top_bar_status_and_search`` (current home) and
-        ``_setup_bottom_bar`` (legacy callers / tests). Either path produces
-        the same button instance.
-        """
-        if getattr(self, "btn_launch", None) is not None:
-            return
-        btn = QPushButton("Select a Model")
-        btn.setEnabled(False)
-        # Top-bar height: align with search input rather than the old 50px
-        # tile-overlapping bottom bar. Width caps the button so a long model
-        # name (e.g. "Launch golf_swing_pendulum >") doesn't push the rest of
-        # the top bar offscreen.
-        btn.setFixedHeight(32)
-        btn.setMinimumWidth(180)
-        btn.setMaximumWidth(280)
-        btn.setFont(get_display_font(size=10, weight=Weights.BOLD))
-        btn.setProperty("class", "launch-ready")
-        _style = btn.style()
-        if _style:
-            _style.polish(btn)
-        btn.clicked.connect(self.launch_simulation)
-        btn.setCursor(Qt.CursorShape.PointingHandCursor)
-        btn.hide()  # Hidden per user request; models are launched via tiles.
-        self.btn_launch = btn
-
-    def _setup_top_bar_config_checkboxes(self, top_bar: QHBoxLayout) -> None:
-        """Create config checkboxes and layout controls, adding them to top bar."""
-        from PyQt6.QtCore import QSettings
-
-        settings = QSettings("UpstreamDrift", "Launcher")
-
-        self.chk_live = QCheckBox("Live Viz")
-        self.chk_live.setChecked(settings.value("chk_live", True, type=bool))
-
-        self.chk_gpu = QCheckBox("GPU")
-        self.chk_gpu.setChecked(settings.value("chk_gpu", False, type=bool))
-
-        self.chk_docker = QCheckBox("Docker")
-        docker_default = getattr(self, "docker_available", False)
-        self.chk_docker.setChecked(
-            settings.value("chk_docker", docker_default, type=bool)
-        )
-        self.chk_docker.stateChanged.connect(self._on_docker_mode_changed)
-
-        self.chk_wsl = QCheckBox("WSL")
-        self.chk_wsl.setChecked(settings.value("chk_wsl", False, type=bool))
-        self.chk_wsl.stateChanged.connect(self._on_wsl_mode_changed)
-
-        self.chk_windows = QCheckBox("Windows")
-        self.chk_windows.setChecked(
-            not self.chk_docker.isChecked() and not self.chk_wsl.isChecked()
-        )
-        self.chk_windows.stateChanged.connect(self._on_windows_mode_changed)
 
         # Layout controls were moved to the View menu per user request.
 
-    def _setup_top_bar_action_buttons(self, top_bar: QHBoxLayout) -> None:
-        """Add Help, Settings, and AI Assistant buttons to top bar."""
         # Action buttons were moved to the left sidebar per user request.
-
-    def _register_top_bar_tooltips(self) -> None:
-        """Register enhanced tooltips for configuration checkboxes."""
-        from src.launchers.launcher_constants import HELP_SYSTEM_AVAILABLE
-
-        if not HELP_SYSTEM_AVAILABLE:
-            return
-
-        from src.shared.python.gui_pkg.help_system import TooltipManager
-
-        TooltipManager.register_tooltip(
-            self.chk_live,
-            "Live Visualization",
-            "Enable real-time 3D visualization during simulation.",
-            "visualization",
-        )
-        TooltipManager.register_tooltip(
-            self.chk_gpu,
-            "GPU Acceleration",
-            "Use GPU for physics computation when available.",
-            "engine_selection",
-        )
-        TooltipManager.register_tooltip(
-            self.chk_docker,
-            "Docker container runtime",
-            "Run physics engines inside the upstream-drift:engine Linux "
-            "container. Full Drake/Pinocchio support; requires Docker "
-            "installed and the image built (Settings → Docker Image).",
-            "engine_selection",
-        )
-        TooltipManager.register_tooltip(
-            self.chk_wsl,
-            "WSL2 Ubuntu runtime",
-            "Run physics engines in your WSL2 Ubuntu user environment. "
-            "Same Linux wheels as Docker mode but no container layer — "
-            "faster file I/O and easier interactive debugging.",
-            "engine_selection",
-        )
-        TooltipManager.register_tooltip(
-            self.chk_windows,
-            "Windows Native runtime",
-            "Run physics engines natively on the local Windows host system.",
-            "engine_selection",
-        )
 
     # ---- View-mode + zoom controls --------------------------------------
 
     _ZOOM_SLIDER_STEPS = 100  # slider integer range -> [MIN, MAX] tile_scale
-
-    def _slider_to_scale(self, value: int) -> float:
-        """Map a slider integer ``value`` to a tile_scale float."""
-        v = max(0, min(self._ZOOM_SLIDER_STEPS, int(value)))
-        frac = v / float(self._ZOOM_SLIDER_STEPS)
-        return TILE_SCALE_MIN + (TILE_SCALE_MAX - TILE_SCALE_MIN) * frac
-
-    def _scale_to_slider(self, scale: float) -> int:
-        """Map a tile_scale float back to slider integer steps."""
-        s = max(TILE_SCALE_MIN, min(TILE_SCALE_MAX, float(scale)))
-        frac = (s - TILE_SCALE_MIN) / (TILE_SCALE_MAX - TILE_SCALE_MIN)
-        return int(round(frac * self._ZOOM_SLIDER_STEPS))
-
-    def _setup_view_mode_and_zoom(self, top_bar: QHBoxLayout) -> None:
-        """Add discrete view-mode dropdown and a compact, elegant zoom slider to top bar."""
-        if top_bar is None:
-            raise ValueError("top_bar must be provided")
-
-        self._top_viewmode_actions: dict[Any, QAction] = {}
-
-        self.view_mode_combo = QComboBox(self.launcher)
-        self.view_mode_combo.addItem("Tile Large", ViewMode.LARGE)
-        self.view_mode_combo.addItem("Tile Medium", ViewMode.MEDIUM)
-        self.view_mode_combo.addItem("Tile Small", ViewMode.SMALL)
-        self.view_mode_combo.addItem("List Large", ViewMode.LIST_LARGE)
-        self.view_mode_combo.addItem("List Small", ViewMode.LIST_SMALL)
-        self.view_mode_combo.setCurrentIndex(3)  # List Large default
-        self.view_mode_combo.setToolTip("Choose how the model tiles are arranged")
-        self.view_mode_combo.setAccessibleName("View mode")
-        self.view_mode_combo.currentIndexChanged.connect(self._on_view_mode_changed)
-
-        self.view_mode_combo.setStyleSheet("""
-            QComboBox {
-                background: #1e1e1e;
-                border: 1px solid #3a3a3a;
-                border-radius: 4px;
-                padding: 4px 8px;
-                color: #cccccc;
-                font-size: 11px;
-                min-width: 100px;
-            }
-            QComboBox:hover {
-                background: #2a2a2a;
-                border-color: #555555;
-                color: #ffffff;
-            }
-            QComboBox QAbstractItemView {
-                background: #1e1e1e;
-                border: 1px solid #3a3a3a;
-                color: #cccccc;
-                selection-background-color: #2a2a2a;
-                selection-color: #ffffff;
-            }
-        """)
-        # Removed view selector dropdown from main screen top bar per user request.
-        # It is kept instantiated and hidden to preserve compatibility with menu actions and settings.
-        self.view_mode_combo.hide()
-
-        self.zoom_slider = QSlider(Qt.Orientation.Horizontal, self.launcher)
-        self.zoom_slider.setRange(0, self._ZOOM_SLIDER_STEPS)
-        self.zoom_slider.setMinimumWidth(140)
-        from PyQt6.QtWidgets import QSizePolicy
-
-        self.zoom_slider.setSizePolicy(
-            QSizePolicy.Policy.MinimumExpanding,
-            self.zoom_slider.sizePolicy().verticalPolicy(),
-        )
-
-        self.zoom_slider.setStyleSheet("""
-            QSlider::groove:horizontal {
-                border: 1px solid #3a3a3a;
-                height: 4px;
-                background: #1a1a1a;
-                margin: 0px;
-                border-radius: 2px;
-            }
-            QSlider::handle:horizontal {
-                background: #888888;
-                border: 1px solid #555555;
-                width: 10px;
-                height: 10px;
-                margin: -3px 0;
-                border-radius: 5px;
-            }
-            QSlider::handle:horizontal:hover {
-                background: #007acc;
-                border-color: #0098ff;
-            }
-        """)
-        self.zoom_slider.setToolTip("Adjust the size of the model tiles")
-        self.zoom_slider.setAccessibleName("Tile zoom")
-        self.zoom_slider.setAccessibleDescription(_build_zoom_accessible_description())
-        self.zoom_slider.setFocusPolicy(Qt.FocusPolicy.StrongFocus)
-
-        from src.launchers.launcher_constants import TILE_SCALE_DEFAULT
-
-        initial_scale = TILE_SCALE_DEFAULT
-        lm = getattr(self, "layout_manager", None)
-        if lm is not None and hasattr(lm, "tile_scale"):
-            initial_scale = float(lm.tile_scale)
-        self.zoom_slider.setValue(self._scale_to_slider(initial_scale))
-        self.zoom_slider.valueChanged.connect(self._on_zoom_slider_changed)
-        self.zoom_slider.hide()
-
-        self.lbl_zoom_pct = QLabel(f"{int(round(initial_scale * 100))}%", self.launcher)
-        self.lbl_zoom_pct.setToolTip("Current tile size as a percentage of base")
-        self.lbl_zoom_pct.setStyleSheet(
-            "font-size: 10px; color: #888888; font-family: monospace;"
-        )
-        self.lbl_zoom_pct.hide()
-
-        # Ctrl+= / Ctrl+- shortcuts adjust zoom by one step (~1.75% scale).
-        sc_in = QShortcut(QKeySequence("Ctrl+="), self.launcher)
-        sc_in.activated.connect(lambda: self._nudge_zoom(+5))
-        sc_in_alt = QShortcut(QKeySequence("Ctrl++"), self.launcher)
-        sc_in_alt.activated.connect(lambda: self._nudge_zoom(+5))
-        sc_out = QShortcut(QKeySequence("Ctrl+-"), self.launcher)
-        sc_out.activated.connect(lambda: self._nudge_zoom(-5))
-
-    def _nudge_zoom(self, delta_steps: int) -> None:
-        """Adjust the zoom slider by ``delta_steps`` integer ticks."""
-        slider = self.zoom_slider
-        if slider is None:
-            return
-        slider.setValue(slider.value() + delta_steps)
-
-    def wheelEvent(self, event: Any) -> None:  # noqa: N802
-        """Ctrl+scroll wheel adjusts the zoom slider."""
-        from PyQt6.QtCore import Qt as _Qt
-
-        if event.modifiers() & _Qt.KeyboardModifier.ControlModifier:
-            delta = event.angleDelta().y()
-            if delta != 0:
-                steps = 5 if delta > 0 else -5
-                self._nudge_zoom(steps)
-                event.accept()
-                return
-        super().wheelEvent(event)  # type: ignore[misc]
-
-    def _on_view_mode_changed(self, index: int) -> None:
-        """Apply the selected view mode to the layout manager + grid."""
-        combo = self.view_mode_combo
-        if combo is None:
-            return
-        mode = combo.itemData(index)
-        is_vm = isinstance(mode, ViewMode) or (
-            hasattr(mode, "__class__") and mode.__class__.__name__ == "ViewMode"
-        )
-        if not is_vm:
-            return
-        if not isinstance(mode, ViewMode):
-            try:
-                mode = ViewMode(int(mode))
-            except (ValueError, TypeError):
-                with contextlib.suppress(AttributeError, KeyError):
-                    mode = ViewMode[mode.name]
-        self._apply_view_mode(mode, sync_combo=False)
-
-    def _set_view_mode_from_menu(self, mode: ViewMode) -> None:
-        """Drive the view-mode change from the View menu's submenu."""
-        self._apply_view_mode(mode, sync_combo=True)
-
-    def _apply_view_mode(self, mode: ViewMode, *, sync_combo: bool) -> None:
-        """Single source of truth for changing tile layout mode.
-
-        Keeps the menubar action group, the top-bar dropdown, the zoom
-        slider, and the grid in sync regardless of which surface
-        triggered the change.
-        """
-        lm = self.layout_manager
-        if lm is None:
-            return
-        lm.set_view_mode(mode)
-        # Sync menu action checkmarks regardless.
-        actions = self._viewmode_actions
-        if actions and mode in actions and not actions[mode].isChecked():
-            actions[mode].setChecked(True)
-        # Sync top-bar dropdown menu action checkmarks regardless.
-        top_actions = self._top_viewmode_actions
-        if top_actions and mode in top_actions and not top_actions[mode].isChecked():
-            top_actions[mode].setChecked(True)
-        # Update zoom slider/label to reflect the mode's default scale.
-        if True:
-            self.zoom_slider.blockSignals(True)
-            self.zoom_slider.setValue(self._scale_to_slider(lm.tile_scale))
-            self.zoom_slider.blockSignals(False)
-        if True:
-            self.lbl_zoom_pct.setText(f"{int(round(lm.tile_scale * 100))}%")
-        if True:
-            lm.rebuild_grid(self.grid_layout)
-        if True:
-            self._save_layout()
-
-    def _on_zoom_slider_changed(self, value: int) -> None:
-        """Live-resize all model cards to match the new slider position."""
-        lm = self.layout_manager
-        scale = self._slider_to_scale(value)
-        if True:
-            self.lbl_zoom_pct.setText(f"{int(round(scale * 100))}%")
-        if lm is None:
-            return
-        lm.set_tile_scale(scale)
-        if True:
-            self._rebuild_grid()
-        if True:
-            self._save_layout()
-
-    def _setup_top_bar(self) -> QHBoxLayout:
-        """Set up the top tool bar."""
-        top_bar = QHBoxLayout()
-
-        # Modern toggles for the sidebars (left nav and right sidekick)
-        self.btn_toggle_left_sidebar = QToolButton(self.launcher)
-        try:
-            from src.shared.python.theme.icon_utils import IconColorizer
-
-            self.btn_toggle_left_sidebar.setIcon(
-                IconColorizer.get_icon("menu", "#cccccc")
-            )
-        except ImportError:
-            self.btn_toggle_left_sidebar.setText("☰")
-        self.btn_toggle_left_sidebar.setToolTip("Toggle Navigation Sidebar")
-        self.btn_toggle_left_sidebar.setCheckable(True)
-        self.btn_toggle_left_sidebar.setChecked(True)
-        self.btn_toggle_left_sidebar.setCursor(Qt.CursorShape.PointingHandCursor)
-        self.btn_toggle_left_sidebar.setStyleSheet(
-            "QToolButton { background: transparent; padding: 4px 8px; border-radius: 4px; } QToolButton:hover { background: #2a2a2a; }"
-        )
-        self.btn_toggle_left_sidebar.clicked.connect(self._toggle_left_sidebar)
-        top_bar.addWidget(self.btn_toggle_left_sidebar)
-
-        self._setup_top_bar_status_and_search(top_bar)
-        self._setup_view_mode_and_zoom(top_bar)
-        self._setup_top_bar_config_checkboxes(top_bar)
-        self._setup_top_bar_action_buttons(top_bar)
-
-        self.btn_toggle_right_sidebar = QToolButton(self.launcher)
-        try:
-            from src.shared.python.theme.icon_utils import IconColorizer
-
-            self.btn_toggle_right_sidebar.setIcon(
-                IconColorizer.get_icon("chat", "#cccccc")
-            )
-        except ImportError:
-            self.btn_toggle_right_sidebar.setText("💬")
-        self.btn_toggle_right_sidebar.setToolTip("Toggle Sidekick Chat")
-        self.btn_toggle_right_sidebar.setCheckable(True)
-        self.btn_toggle_right_sidebar.setChecked(True)
-        self.btn_toggle_right_sidebar.setCursor(Qt.CursorShape.PointingHandCursor)
-        self.btn_toggle_right_sidebar.setStyleSheet(
-            "QToolButton { background: transparent; padding: 4px 8px; border-radius: 4px; } QToolButton:hover { background: #2a2a2a; }"
-        )
-        self.btn_toggle_right_sidebar.clicked.connect(self._toggle_sidekick)
-        top_bar.addWidget(self.btn_toggle_right_sidebar)
-
-        # Context Help Dock
-        self._setup_context_help()
-
-        # Register enhanced tooltips
-        self._register_top_bar_tooltips()
-
-        return top_bar
 
     def _setup_grid_area(self, layout: QVBoxLayout) -> None:
         """Set up the scrollable grid area."""
@@ -2181,7 +902,6 @@ class UISetupManager:
 
     def _kill_process_by_name(self, name: str) -> None:
         """Kill a running process by its name."""
-        import contextlib
 
         with self.process_manager._process_lock:
             proc = self.running_processes.get(name)
